+++ html/pfappserver/root/src/views/Configuration/pki/revokedCerts/_composables/useCollection.js
@@ -19,7 +19,7 @@ export { useRouter } from '../_router'
 export { useStore } from '../_store'
 
 import { pfSearchConditionType as conditionType } from '@/globals/pfSearch'
-import makeSearch from '@/views/Configuration/_store/factory/search'
+import makeSearch from '@/store/factory/search'
 import api from '../_api'
 import { revokeReasons } from '../../config'
 export const useSearch = makeSearch('pkiRevokedCerts', {
