+++ html/pfappserver/root/src/views/Configuration/pkiProviders/_composables/useCollection.js
@@ -49,7 +49,7 @@ export { useRouter } from '../_router'
 export { useStore } from '../_store'
 
 import { pfSearchConditionType as conditionType } from '@/globals/pfSearch'
-import makeSearch from '@/views/Configuration/_store/factory/search'
+import makeSearch from '@/store/factory/search'
 import api from '../_api'
 export const useSearch = makeSearch('pkiProviders', {
   api,
