+++ html/pfappserver/root/src/views/Configuration/scanEngines/_composables/useCollection.js
@@ -51,7 +51,7 @@ export { useRouter } from '../_router'
 export { useStore } from '../_store'
 
 import { pfSearchConditionType as conditionType } from '@/globals/pfSearch'
-import makeSearch from '@/views/Configuration/_store/factory/search'
+import makeSearch from '@/store/factory/search'
 import api from '../_api'
 export const useSearch = makeSearch('scanEngines', {
   api,
